bgituQuizzAdminTest/requirements.txt
asgiref
contourpy
cycler
diff-match-patch
Django==5.2.2
django-import-export
django-unfold
et_xmlfile
fonttools
kiwisolver
matplotlib
numpy<2.0
openpyxl
packaging
pandas
pillow
psycopg2-binary
pyparsing
python-dateutil
six
sqlparse
tablib
django-ckeditor